+++ Project2-Stream-Compaction/stream_compaction/efficient.cu
@@ -18,17 +18,17 @@ namespace StreamCompaction {
 			return timer;
 		}
 		/*2 scan phases, see (link)[https://developer.nvidia.com/gpugems/GPUGems3/gpugems3_ch39.html] for more details*/
-		__global__ void reduce_parallel(unsigned long int n, int *data, int d) {
-			unsigned long int tmp_d = 1 << (d + 1);
-			unsigned long int index = (blockDim.x * blockIdx.x + threadIdx.x)*tmp_d;
+		__global__ void reduce_parallel(unsigned long long int n, int *data, int d) {
+			unsigned long long int tmp_d = 1 << (d + 1);
+			unsigned long long int index = (blockDim.x * blockIdx.x + threadIdx.x)*tmp_d;
 			if (index >= n )
 				return;
 			data[index + tmp_d - 1] += data[index + (tmp_d>>1) - 1];
 		}
 
-		__global__ void downsweep_parallel(unsigned long int n, int *data, int d) {
-			unsigned long int new_d = 1<<(d + 1);
-			unsigned long int index = (blockDim.x * blockIdx.x + threadIdx.x)*new_d;
+		__global__ void downsweep_parallel(unsigned long long int n, int *data, int d) {
+			unsigned long long int new_d = 1<<(d + 1);
+			unsigned long long int index = (blockDim.x * blockIdx.x + threadIdx.x)*new_d;
 			if (index >= n)
 				return;
 			int t = data[index + (new_d>>1) - 1];
@@ -38,12 +38,12 @@ namespace StreamCompaction {
 		/**
 		 * Performs prefix-sum (aka scan) on idata, storing the result into odata.
 		 */
-		void scan(unsigned long int n, int *odata, const int *idata) {
+		void scan(unsigned long long int n, int *odata, const int *idata) {
 			timer().startGpuTimer();
 			// allocate pointers to memory and copy data over
 			int *dev_odata;
-			unsigned long int blocks = 0;
-			unsigned long int closest_pow2 = 1<<ilog2ceil(n);
+			unsigned long long int blocks = 0;
+			unsigned long long int closest_pow2 = 1<<ilog2ceil(n);
 			cudaMalloc((void**)&dev_odata, closest_pow2 * sizeof(int));
 			checkCUDAErrorWithLine("malloc failed!");
 			// we can use dev_odata to hold idata (because we do the scan inplace)
@@ -73,10 +73,10 @@ namespace StreamCompaction {
 		}
 
 		/*Copy of scan but only works with cuda pointers*/
-		void dev_scan(unsigned long int n, int *dev_odata) {
+		void dev_scan(unsigned long long int n, int *dev_odata) {
 			// allocate pointers to memory and copy data over
-			unsigned long int blocks = 0;
-			unsigned long int closest_pow2 = 1 << ilog2ceil(n);
+			unsigned long long int blocks = 0;
+			unsigned long long int closest_pow2 = 1 << ilog2ceil(n);
 			// reduce phase
 			// so we dont need to do the last round of computation because we zero it anyway
 			for (int d = 0; d <= ilog2ceil(closest_pow2) - 2; d++) {
@@ -104,11 +104,11 @@ namespace StreamCompaction {
 		 * @param idata  The array of elements to compact.
 		 * @returns      The number of elements remaining after compaction.
 		 */
-		unsigned long int compact(unsigned long int n, int *odata, const int *idata) {
+		unsigned long long int compact(unsigned long long int n, int *odata, const int *idata) {
 			timer().startGpuTimer();
-			unsigned long int closest_pow2 = 1<<ilog2ceil(n);
+			unsigned long long int closest_pow2 = 1<<ilog2ceil(n);
 			int *dev_idata, *dev_odata, *dev_mask, *dev_indices;
-			unsigned long int blocks = ceil((closest_pow2 + block_size - 1) / block_size);
+			unsigned long long int blocks = ceil((closest_pow2 + block_size - 1) / block_size);
 			cudaMalloc((void**)&dev_idata, closest_pow2 * sizeof(int));
 			cudaMalloc((void**)&dev_odata, closest_pow2 * sizeof(int));
 			cudaMalloc((void**)&dev_mask, closest_pow2 * sizeof(int));
@@ -149,30 +149,30 @@ namespace StreamCompaction {
 		#define LOG_NUM_BANKS 4
 		#define CONFLICT_FREE_OFFSET(n) \
 			((n) >> NUM_BANKS + (n) >> (2 * LOG_NUM_BANKS))
-		__global__ void dev_scan(unsigned long int n, int *dev_odata, int *dev_idata, int *dev_block_sum)
+		__global__ void dev_scan(unsigned long long int n, int *dev_odata, int *dev_idata, int *dev_block_sum)
 		{
 			/* Extriemly heavly based on https://developer.nvidia.com/gpugems/GPUGems3/gpugems3_ch39.html
 			   Main change is to use block_size instead of n and then put the sums of each block into a new array and scaning the array
 			*/
 			// Declare Share Memory
 			__shared__ int temp[block_size + NUM_BANKS];
-			unsigned long int thid = threadIdx.x;
-			unsigned long int bid = blockIdx.x;
-			unsigned long int scan_offset = bid * block_size;
+			unsigned long long int thid = threadIdx.x;
+			unsigned long long int bid = blockIdx.x;
+			unsigned long long int scan_offset = bid * block_size;
 			int offset = 1; // to make this an exclusive scan
-			unsigned long int ai = thid<<1;
-			unsigned long int bi = ai + 1;
-			unsigned long int bankOffsetA = CONFLICT_FREE_OFFSET(ai);
-			unsigned long int bankOffsetB = CONFLICT_FREE_OFFSET(bi);
+			unsigned long long int ai = thid<<1;
+			unsigned long long int bi = ai + 1;
+			unsigned long long int bankOffsetA = CONFLICT_FREE_OFFSET(ai);
+			unsigned long long int bankOffsetB = CONFLICT_FREE_OFFSET(bi);
 			temp[ai + bankOffsetA] = dev_idata[ai + scan_offset];
 			temp[bi + bankOffsetB] = dev_idata[bi + scan_offset];
-			for (unsigned long int d = block_size >> 1; d > 0; d >>= 1)                    // build sum in place up the tree
+			for (unsigned long long int d = block_size >> 1; d > 0; d >>= 1)                    // build sum in place up the tree
 			{
 				__syncthreads();
 				if (thid < d)
 				{
-					unsigned long int ai = offset * ((thid<<1) + 1) - 1;
-					unsigned long int bi = offset * ((thid << 1) + 2) - 1;
+					unsigned long long int ai = offset * ((thid<<1) + 1) - 1;
+					unsigned long long int bi = offset * ((thid << 1) + 2) - 1;
 					ai += CONFLICT_FREE_OFFSET(ai);
 					bi += CONFLICT_FREE_OFFSET(bi);
 					temp[bi] += temp[ai];
@@ -187,14 +187,14 @@ namespace StreamCompaction {
 				temp[block_size - 1 + CONFLICT_FREE_OFFSET(block_size - 1)] = 0; 
 			}
 			// downsweep
-			for (unsigned long int d = 1; d < block_size; d<<=1) // traverse down tree & build scan
+			for (unsigned long long int d = 1; d < block_size; d<<=1) // traverse down tree & build scan
 			{
 				offset >>= 1;
 				__syncthreads();
 				if (thid < d)
 				{
-					unsigned long int ai = offset * ((thid << 1) + 1) - 1;
-					unsigned long int bi = offset * ((thid << 1) + 2) - 1;
+					unsigned long long int ai = offset * ((thid << 1) + 1) - 1;
+					unsigned long long int bi = offset * ((thid << 1) + 2) - 1;
 					ai += CONFLICT_FREE_OFFSET(ai);
 					bi += CONFLICT_FREE_OFFSET(bi);
 					float t = temp[ai];
@@ -206,22 +206,22 @@ namespace StreamCompaction {
 			dev_odata[ai + scan_offset] = temp[ai + bankOffsetA];
 			dev_odata[bi + scan_offset] = temp[bi + bankOffsetB];
 		}
-		__global__ void add_offset(unsigned long int n, int *data, int *dev_block_offset) {
-			unsigned long int bid = blockIdx.x;
-			unsigned long int index = blockDim.x * bid + threadIdx.x;
+		__global__ void add_offset(unsigned long long int n, int *data, int *dev_block_offset) {
+			unsigned long long int bid = blockIdx.x;
+			unsigned long long int index = blockDim.x * bid + threadIdx.x;
 			if (index >= n)
 				return;
 			// add value to current section
 			if (bid != 0) // to save a bunch of useless reads and write
 				data[index] += dev_block_offset[bid];
 		}
-		void scan(unsigned long int n, int *odata, int *idata) {
+		void scan(unsigned long long int n, int *odata, int *idata) {
 			timer().startGpuTimer();
 			// allocate pointers to memory and copy data over
 			int *dev_odata, *dev_idata, *dev_block_sum, *dev_block_offset;
 			int *block_sum, *block_offset;
-			unsigned long int closest_pow2 = 1 << ilog2ceil(n);
-			unsigned long int blocks = ceil((closest_pow2 + block_size - 1) / block_size);
+			unsigned long long int closest_pow2 = 1 << ilog2ceil(n);
+			unsigned long long int blocks = ceil((closest_pow2 + block_size - 1) / block_size);
 			// allocate buffers
 			cudaMalloc((void**)&dev_odata, closest_pow2 * sizeof(int));
 			cudaMalloc((void**)&dev_idata, closest_pow2 * sizeof(int));
@@ -259,13 +259,13 @@ namespace StreamCompaction {
 			delete[] block_offset;
 			timer().endGpuTimer();
 		}
-		void dev_scan(unsigned long int n, int *dev_odata, int *dev_idata) {
+		void dev_scan(unsigned long long int n, int *dev_odata, int *dev_idata) {
 			timer().startGpuTimer();
 			// allocate pointers to memory and copy data over
 			int *dev_block_sum, *dev_block_offset;
 			int *block_sum, *block_offset;
-			unsigned long int closest_pow2 = 1 << ilog2ceil(n);
-			unsigned long int blocks = ceil((closest_pow2 + block_size - 1) / block_size);
+			unsigned long long int closest_pow2 = 1 << ilog2ceil(n);
+			unsigned long long int blocks = ceil((closest_pow2 + block_size - 1) / block_size);
 			// allocate buffers
 			// allocate block global buffers
 			cudaMalloc((void**)&dev_block_sum, blocks * sizeof(int)); // each block gets 1 number to fill in
