+++ Project2-Stream-Compaction/stream_compaction/radix.h
@@ -5,6 +5,6 @@
 namespace Sorting {
 	namespace Radix {
 		StreamCompaction::Common::PerformanceTimer& timer();
-		void sort(unsigned long int n, int *odata, int *idata, int max_value);
+		void sort(unsigned long long int n, int *odata, int *idata, int max_value);
 	}
 }
